feat: create token rs
